Spatial Sharing test update for low variance
