+++ src/runtime_src/core/tools/common/TestRunner.h
@@ -46,8 +46,6 @@ class TestRunner : public JSONConfigurable {
     std::vector<std::string> findDependencies( const std::string& test_path,
                       const std::string& ps_kernel_name);
     int validate_binary_file(const std::string& binaryfile);
-    void init_instr_buf(xrt::bo &bo_instr, const std::string& dpu_file);
-    size_t get_instr_size(const std::string& dpu_file);
     void result_in_range(double value, double threshold, boost::property_tree::ptree& ptTest);
     void set_threshold(const std::shared_ptr<xrt_core::device>& dev, boost::property_tree::ptree& ptTest);
     double get_threshold(){return m_threshold;};
